Added new component for a message
