+++ web-components/src/index.js
@@ -4,3 +4,4 @@ import './components/AppHeader';
 import './components/FormInput';
 import './components/MessageForm';
 import './components/ChatsList';
+import './components/MessageContainer';
